pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hsa-dpo"
version = "1.0.0"
description = "Detecting and Mitigating Hallucination in Large Vision Language Models via Fine-Grained AI Feedback"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    # Core ML frameworks - keep exact versions for compatibility
    "torch==2.0.1",
    "torchvision==0.15.2",
    "transformers==4.31.0",
    "tokenizers>=0.12.1,<0.14",
    "sentencepiece==0.1.99",
    "accelerate==0.22.0",
    "peft==0.4.0",
    "bitsandbytes==0.41.0",
    "deepspeed==0.9.5",
    
    # LLaVA dependencies (LLaVA code is included in hsa_dpo/models/llava-v1_5/)
    "einops==0.6.1",
    "einops-exts==0.0.4",
    "timm==0.6.13",
    
    # Data processing
    "numpy==1.26.1",
    "Pillow",
    "datasets",
    "jsonlines",
    "scikit-learn==1.2.2",
    
    # Training utilities
    "tensorboard",
    "wandb",
    
    # Additional required utilities
    "shortuuid",
    "requests",
    "pydantic<2,>=1",
    "fairscale==0.4.4",
]

[project.optional-dependencies]
flash-attn = ["flash-attn"]
dev = ["pytest", "black", "flake8"]

[project.urls]

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]